+++ trunk/Extensions/QueryAnalyzer/QueryAnalyzer_ADONET_RegressionTests/DbAggregators_COUNTDISTINCT.cs
@@ -35,7 +35,7 @@ namespace QueryAnalyzer_ADONET_RegressionTests
                         while (reader.Read())
                         {
                             int id = reader.GetInt32(0);
-                            int count = reader.GetInt32(1);
+                            int count = (int)reader.GetInt64(1);
                             if (!expected.ContainsKey(id))
                             {
                                 throw new Exception("id returned from countdistinct group by is invalid.");
@@ -95,7 +95,7 @@ namespace QueryAnalyzer_ADONET_RegressionTests
                         while (reader.Read())
                         {
                             int id = reader.GetInt32(0);
-                            int count = reader.GetInt32(1);
+                            int count = (int)reader.GetInt64(1);
                             if (!expected.ContainsKey(id))
                             {
                                 throw new Exception("id returned from countdistinct group by is invalid.");
